+++ mkl_fft/interfaces/scipy_fft.py
@@ -24,4 +24,4 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from .._scipy_fft_backend import *
+from .._scipy_fft import *
